requirements.txt
torch
torchvision
numpy
opencv-python
tqdm
scikit-learn
pillow
transformers
peft
accelerate